requirements.txt
flask
pypdf
langchain>=0.2.0
langchain-classic
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
tiktoken
